# contracts/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "ContentType": {
        "enum": [
          "main_text",
          "author_note",
          "editor_note",
          "footnote"
        ],
        "title": "ContentType",
        "type": "string"
      },
      "ContextItem": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          },
          "is_target": {
            "title": "Is Target",
            "type": "boolean"
          }
        },
        "required": [
          "evidence",
          "is_target"
        ],
        "title": "ContextItem",
        "type": "object"
      },
      "ContextResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContextItem"
            },
            "title": "Items",
            "type": "array"
          },
          "target_evidence_id": {
            "title": "Target Evidence Id",
            "type": "string"
          }
        },
        "required": [
          "target_evidence_id",
          "items"
        ],
        "title": "ContextResponse",
        "type": "object"
      },
...
      "DatePrecision": {
        "enum": [
          "day",
          "month",
          "year",
          "range",
          "approximate",
          "disputed",
          "unknown"
        ],
        "title": "DatePrecision",
        "type": "string"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "Evidence": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "corpus_name": {
            "title": "Corpus Name",
            "type": "string"
          },
          "date_precision": {
            "$ref": "#/components/schemas/DatePrecision"
          },
          "edition_label": {
            "title": "Edition Label",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "exact_match_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exact Match Count"
          },
          "match_offsets": {
            "items": {
              "type": "integer"
            },
            "title": "Match Offsets",
            "type": "array"
          },
          "match_type": {
            "title": "Match Type",
            "type": "string"
          },
          "next_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Evidence Id"
          },
          "pdf_pages": {
            "items": {
              "type": "integer"
            },
            "title": "Pdf Pages",
            "type": "array"
          },
          "prev_evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Evidence Id"
          },
          "printed_pages": {
            "items": {
              "type": "string"
            },
            "title": "Printed Pages",
            "type": "array"
          },
          "rank_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Rank Reasons",
            "type": "array"
          },
          "support_label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupportLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "verified_text": {
            "minLength": 1,
            "title": "Verified Text",
            "type": "string"
          },
          "volume_no": {
            "minimum": 1.0,
            "title": "Volume No",
            "type": "integer"
          },
          "work_date_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Date End"
          },
          "work_date_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Date Start"
          },
          "work_title": {
            "title": "Work Title",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "verified_text",
          "content_type",
          "author",
          "work_title",
          "corpus_name",
          "edition_label",
          "volume_no",
          "date_precision",
          "match_type"
        ],
        "title": "Evidence",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PdfLocationResponse": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "coordinates": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinates"
          },
          "end_pdf_page": {
            "minimum": 1.0,
            "title": "End Pdf Page",
            "type": "integer"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "printed_pages": {
            "items": {
              "type": "string"
            },
            "title": "Printed Pages",
            "type": "array"
          },
          "start_pdf_page": {
            "minimum": 1.0,
            "title": "Start Pdf Page",
            "type": "integer"
          },
          "viewer_url": {
            "title": "Viewer Url",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "asset_id",
          "viewer_url",
          "start_pdf_page",
          "end_pdf_page",
          "printed_pages"
        ],
        "title": "PdfLocationResponse",
        "type": "object"
      },
...
      "ScopeNode": {
        "additionalProperties": false,
        "properties": {
          "children": {
            "items": {
              "$ref": "#/components/schemas/ScopeNode"
            },
            "title": "Children",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "node_type": {
            "enum": [
              "corpus",
              "edition",
              "volume",
              "work",
              "author"
            ],
            "title": "Node Type",
            "type": "string"
          },
          "release_status": {
            "title": "Release Status",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "node_type",
          "label",
          "release_status"
        ],
        "title": "ScopeNode",
        "type": "object"
      },
      "ScopeTreeResponse": {
        "additionalProperties": false,
        "properties": {
          "corpus_id": {
            "title": "Corpus Id",
            "type": "string"
          },
          "data_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Version"
          },
          "roots": {
            "items": {
              "$ref": "#/components/schemas/ScopeNode"
            },
            "title": "Roots",
            "type": "array"
          }
        },
        "required": [
          "corpus_id"
        ],
        "title": "ScopeTreeResponse",
        "type": "object"
      },
...
      "SupportLabel": {
        "enum": [
          "direct",
          "indirect",
          "counter",
          "context_only",
          "irrelevant"
        ],
        "title": "SupportLabel",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Marx–Engels InsightMatch API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/corpora/{corpus_id}/scope-tree": {
      "get": {
        "operationId": "get_corpus_scope_tree",
        "parameters": [
          {
            "in": "path",
            "name": "corpus_id",
            "required": true,
            "schema": {
              "title": "Corpus Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeTreeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scope Tree",
        "tags": [
          "metadata"
        ]
      }
    },
    "/api/v1/evidence/{evidence_id}": {
      "get": {
        "operationId": "get_evidence",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "evidence"
        ]
      }
    },
    "/api/v1/evidence/{evidence_id}/context": {
      "get": {
        "operationId": "get_evidence_context",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 3,
              "minimum": 0,
              "title": "Before",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 3,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Context",
        "tags": [
          "evidence"
        ]
      }
    },
    "/api/v1/evidence/{evidence_id}/pdf-location": {
      "get": {
        "operationId": "get_evidence_pdf_location",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfLocationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pdf Location",
        "tags": [
          "evidence"
        ]
      }
    },
...
  }
}